pipeline interface tests passing
