+++ looper/pipeline_interface.py
@@ -1,5 +1,6 @@
 """ Model the connection between a pipeline and a project or executor. """
 
+import inspect
 import logging
 import os
 import sys
@@ -13,8 +14,9 @@ import yaml
 from .exceptions import \
     InvalidResourceSpecificationException, \
     MissingPipelineConfigurationException
-from peppy import utils
+from peppy import utils, Sample
 from peppy.const import DEFAULT_COMPUTE_RESOURCES_NAME
+from peppy.utils import is_command_callable
 
 
 
@@ -35,56 +37,66 @@ class PipelineInterface(object):
     :type config: str | Mapping
 
     """
+
+    SUBTYPE_MAPPING_SECTION = "sample_subtypes"
+
+
     def __init__(self, config):
+
         if isinstance(config, Mapping):
-            # Unified pipeline_interface.yaml file (protocol mappings
-            # and the actual pipeline interface data)
-            _LOGGER.debug("Creating %s with preparsed data",
-                          self.__class__.__name__)
             self.pipe_iface_file = None
             self.pipe_iface_config = config
-
+            self.pipelines_path = None
         else:
             # More likely old-style, with protocol_mapping in its own file,
             # separate from the actual pipeline interface data
-            _LOGGER.debug("Parsing '%s' for PipelineInterface config data",
-                          config)
+            _LOGGER.debug("Parsing '%s' for %s config data",
+                          config, self.__class__.__name__)
             self.pipe_iface_file = config
             with open(config, 'r') as f:
                 self.pipe_iface_config = yaml.load(f)
+            self.pipelines_path = os.path.dirname(config)
+        
+        for section in ["protocol_mapping", "pipelines"]:
+            assert section in self.pipe_iface_config, \
+                "{} config must declare section '{}'". \
+                    format(self.__class__.__name__, section)
 
         # Ensure that each pipeline path, if provided, is expanded.
-        self._expand_paths()
-
-
-    def __getitem__(self, item):
-        try:
-            return self._select_pipeline(item)
-        except MissingPipelineConfigurationException:
-            raise KeyError("{} is not a known pipeline; known: {}".
-                           format(item, self.pipe_iface_config.keys()))
-
-
-    def __iter__(self):
-        return iter(self.pipe_iface_config.items())
-
-
-    def __repr__(self):
-        source = self.pipe_iface_file or "Mapping"
-        num_pipelines = len(self.pipe_iface_config)
-        pipelines = ", ".join(self.pipe_iface_config.keys())
-        return "{} from {}, with {} pipeline(s): {}".format(
-                self.__class__.__name__, source, num_pipelines, pipelines)
-
-
-    def _expand_paths(self):
-        for pipe_data in self.pipe_iface_config.values():
+        for pipe_data in self.pipe_iface_config["pipelines"].values():
             if "path" in pipe_data:
                 pipe_path = pipe_data["path"]
                 _LOGGER.log(5, "Expanding path: '%s'", pipe_path)
                 pipe_path = utils.expandpath(pipe_path)
                 _LOGGER.log(5, "Expanded: '%s'", pipe_path)
                 pipe_data["path"] = pipe_path
+
+        self.pipe_iface = self.pipe_iface_config["pipelines"]
+
+
+    def __getitem__(self, item):
+        """ Mapping index-like syntax is interpreted as pipeline request. """
+        try:
+            return self.select_pipeline(item)
+        except MissingPipelineConfigurationException:
+            raise KeyError(
+                "{} is not a known pipeline; known: {}".
+                format(item, self.pipe_iface.keys()))
+
+
+    def __iter__(self):
+        """ Iteration is over K-V pairs in 'pipelines' section. """
+        return iter(self.pipe_iface.items())
+
+
+    def __repr__(self):
+        """ String representation """
+        source = self.pipe_iface_file or "Mapping"
+        num_pipelines = len(self.pipe_iface)
+        # TODO: could use 'name' here
+        pipelines = ", ".join(self.pipe_iface.keys())
+        return "{} from {}, with {} pipeline(s): {}".format(
+                self.__class__.__name__, source, num_pipelines, pipelines)
 
 
     @property
@@ -95,7 +107,8 @@ class PipelineInterface(object):
         :return Iterable[str]: names of pipelines about which this
             interface is aware
         """
-        return self.pipe_iface_config.keys()
+        # TODO: could consider keying on name.
+        return self.pipe_iface.keys()
 
 
     @property
@@ -105,7 +118,7 @@ class PipelineInterface(object):
 
         :return Mapping: pipeline interface configuration data
         """
-        return self.pipe_iface_config.values()
+        return self.pipe_iface.values()
 
 
     def choose_resource_package(self, pipeline_name, file_size):
@@ -133,7 +146,7 @@ class PipelineInterface(object):
                              "negative file size: {}".format(file_size))
 
         try:
-            resources = self._select_pipeline(pipeline_name)["resources"]
+            resources = self.select_pipeline(pipeline_name)["resources"]
         except KeyError:
             msg = "No resources for pipeline '{}'".format(pipeline_name)
             if self.pipe_iface_file is not None:
@@ -193,6 +206,60 @@ class PipelineInterface(object):
                 return rp_data
 
 
+    def finalize_pipeline_key_and_paths(self, pipeline_key):
+        """
+        Determine pipeline's full path, arguments, and strict key.
+
+        This handles multiple ways in which to refer to a pipeline (by key)
+        within the mapping that contains the data that defines a
+        PipelineInterface. It also ensures proper handling of the path to the
+        pipeline (i.e., ensuring that it's absolute), and that the text for
+        the arguments are appropriately dealt parsed and passed.
+
+        :param str pipeline_key: the key in the pipeline interface file used
+            for the protocol_mappings section. Previously was the script name.
+        :return (str, str, str): more precise version of input key, along with
+            absolute path for pipeline script, and full script path + options
+
+        """
+
+        # The key may contain extra command-line flags; split key from flags.
+        # The strict key is the script name itself, something like "ATACseq.py"
+        strict_pipeline_key, _, pipeline_key_args = pipeline_key.partition(' ')
+
+        full_pipe_path = \
+                self.get_attribute(strict_pipeline_key, "path")
+        if full_pipe_path:
+            script_path_only = os.path.expanduser(
+                os.path.expandvars(full_pipe_path[0].strip()))
+            if os.path.isdir(script_path_only):
+                script_path_only = os.path.join(script_path_only, pipeline_key)
+            script_path_with_flags = \
+                    "{} {}".format(script_path_only, pipeline_key_args)
+        else:
+            # backwards compatibility w/ v0.5
+            script_path_only = strict_pipeline_key
+            script_path_with_flags = pipeline_key
+
+        # Clear trailing whitespace.
+        script_path_only = script_path_only.rstrip()
+
+        # TODO: determine how to deal with pipeplines_path
+        if not os.path.isabs(script_path_only) and not \
+                is_command_callable(script_path_only):
+            _LOGGER.log(5, "Expanding non-absolute script path: '%s'",
+                        script_path_only)
+            script_path_only = os.path.join(
+                    self.pipelines_path, script_path_only)
+            _LOGGER.log(5, "Absolute script path: '%s'", script_path_only)
+            script_path_with_flags = os.path.join(
+                    self.pipelines_path, script_path_with_flags)
+            _LOGGER.log(5, "Absolute script path with flags: '%s'",
+                        script_path_with_flags)
+
+        return strict_pipeline_key, script_path_only, script_path_with_flags
+
+
     def get_arg_string(self, pipeline_name, sample,
                        submission_folder_path="", **null_replacements):
         """
@@ -225,7 +292,7 @@ class PipelineInterface(object):
         proxies.update(null_replacements)
 
         _LOGGER.debug("Building arguments string")
-        config = self._select_pipeline(pipeline_name)
+        config = self.select_pipeline(pipeline_name)
         argstring = ""
 
         if "arguments" not in config:
@@ -305,7 +372,7 @@ class PipelineInterface(object):
             is returned as a list; this is useful for safe iteration over
             the returned value.
         """
-        config = self._select_pipeline(pipeline_name)
+        config = self.select_pipeline(pipeline_name)
         value = config.get(attribute_key)
         return [value] if isinstance(value, str) and path_as_list else value
 
@@ -321,12 +388,97 @@ class PipelineInterface(object):
             stripping the pipeline's file extension
         :rtype: str: translated name for pipeline
         """
-        config = self._select_pipeline(pipeline)
+        config = self.select_pipeline(pipeline)
         try:
             return config["name"]
         except KeyError:
             _LOGGER.debug("No 'name' for pipeline '{}'".format(pipeline))
             return os.path.splitext(pipeline)[0]
+
+
+    def fetch_pipelines(self, protocol):
+        """
+        Fetch the mapping for a particular protocol, null if unmapped.
+
+        :param str protocol: name/key for the protocol for which to fetch the
+            pipeline(s)
+        :return str | Iterable[str] | NoneType: pipeline(s) to which the given
+            protocol is mapped, otherwise null
+        """
+        protocol_key = utils.alpha_cased(protocol)
+        return self.protomap.get(protocol_key)
+
+
+    def fetch_sample_subtype(
+            self, protocol, strict_pipe_key, full_pipe_path):
+        """
+        Determine the interface and Sample subtype for a protocol and pipeline.
+
+        :param str protocol: name of the relevant protocol
+        :param str strict_pipe_key: key for specific pipeline in a pipeline
+            interface mapping declaration; this must exactly match a key in
+            the PipelineInterface (or the Mapping that represent it)
+        :param str full_pipe_path: (absolute, expanded) path to the
+            pipeline script
+        :return type: Sample subtype to use for jobs for the given protocol,
+            that use the pipeline indicated
+        :raises KeyError: if given a pipeline key that's not mapped in this
+            ProtocolInterface instance's PipelineInterface
+        """
+
+        subtype = None
+
+        this_pipeline_data = self.pipe_iface[strict_pipe_key]
+
+        try:
+            subtypes = this_pipeline_data[self.SUBTYPE_MAPPING_SECTION]
+        except KeyError:
+            _LOGGER.debug("%s from '%s' doesn't define section '%s' "
+                          "for pipeline '%s'",
+                          self.pipe_iface.__class__.__name__, self.source,
+                          self.SUBTYPE_MAPPING_SECTION, strict_pipe_key)
+            # Without a subtypes section, if pipeline module defines a single
+            # Sample subtype, we'll assume that type is to be used when in
+            # this case, when the interface section for this pipeline lacks
+            # an explicit subtypes section specification.
+            subtype_name = None
+        else:
+            if subtypes is None:
+                # Designate lack of need for import attempt and provide
+                # class with name to format message below.
+                subtype = Sample
+                _LOGGER.debug("Null %s subtype(s) section specified for "
+                              "pipeline: '%s'; using base %s type",
+                              subtype.__name__, strict_pipe_key,
+                              subtype.__name__)
+            elif isinstance(subtypes, str):
+                subtype_name = subtypes
+                _LOGGER.debug("Single subtype name for pipeline '%s' "
+                              "in interface from '%s': '%s'", subtype_name,
+                              strict_pipe_key, self.source)
+            else:
+                temp_subtypes = {
+                        utils.alpha_cased(p): st for p, st in subtypes.items()}
+                try:
+                    subtype_name = temp_subtypes[utils.alpha_cased(protocol)]
+                except KeyError:
+                    # Designate lack of need for import attempt and provide
+                    # class with name to format message below.
+                    subtype = Sample
+                    _LOGGER.debug("No %s subtype specified in interface from "
+                                  "'%s': '%s', '%s'; known: %s",
+                                  subtype.__name__, self.source,
+                                  strict_pipe_key, protocol,
+                                  ", ".join(temp_subtypes.keys()))
+
+        # subtype_name is defined if and only if subtype remained null.
+        # The import helper function can return null if the import attempt
+        # fails, so provide the base Sample type as a fallback.
+        subtype = subtype or \
+                  _import_sample_subtype(full_pipe_path, subtype_name) or \
+                  Sample
+        _LOGGER.debug("Using Sample subtype: %s", subtype.__name__)
+        return subtype
 
 
     def uses_looper_args(self, pipeline_name):
@@ -339,11 +491,47 @@ class PipelineInterface(object):
         :return: Whether indicated pipeline accepts looper arguments.
         :rtype: bool
         """
-        config = self._select_pipeline(pipeline_name)
+        config = self.select_pipeline(pipeline_name)
         return "looper_args" in config and config["looper_args"]
 
 
-    def _select_pipeline(self, pipeline_name):
+    @classmethod
+    def _parse_iface_data(cls, pipe_iface_data):
+        """
+        Parse data from mappings to set instance attributes.
+
+        The data that define a ProtocolInterface are a "protocol_mapping"
+        Mapping and a "pipelines" Mapping, which are used to create a
+        ProtocolMapper and a PipelineInterface, representing the configuration
+        data for pipeline(s) from a single location. There are a couple of
+        different ways (file, folder, and eventually, raw Mapping) to provide
+        this data, and this function provides some standardization to how
+        those data are processed, independent of input type/format.
+
+        :param Mapping[str, Mapping] pipe_iface_data: mapping from section
+            name to section data mapping; more specifically, the protocol
+            mappings Mapping and the PipelineInterface mapping
+        :return list[(str, ProtocolMapper | PipelineInterface)]: pairs of
+            attribute name for the ProtocolInterface being created, and the
+            value for that attribute,
+        """
+        assignments = [("protocol_mapping", None, "protomap"),
+                       ("pipelines", PipelineInterface, "pipe_iface")]
+        attribute_values = []
+        for section_name, data_type, attr_name in assignments:
+            try:
+                data = pipe_iface_data[section_name]
+            except KeyError:
+                _LOGGER.error("Error creating %s from data: %s",
+                              cls.__name__, str(pipe_iface_data))
+                raise Exception("PipelineInterface file lacks section: '{}'".
+                                format(section_name))
+            data = data if data_type is None else data_type(data)
+            attribute_values.append((attr_name, data))
+        return attribute_values
+
+
+    def select_pipeline(self, pipeline_name):
         """
         Check to make sure that pipeline has an entry and if so, return it.
 
@@ -356,7 +544,7 @@ class PipelineInterface(object):
         """
         try:
             # For unmapped pipeline, Return empty config instead of None.
-            return self.pipe_iface_config[pipeline_name] or dict()
+            return self.pipe_iface[pipeline_name] or dict()
         except KeyError:
             _LOGGER.error(
                 "Missing pipeline description: %s not found; %d known: %s",
@@ -364,3 +552,124 @@ class PipelineInterface(object):
                 ", ".format(self.pipe_iface_config.keys()))
             # TODO: use defaults or force user to define this?
             raise MissingPipelineConfigurationException(pipeline_name)
+
+
+
+def _import_sample_subtype(pipeline_filepath, subtype_name=None):
+    """
+    Import a particular Sample subclass from a Python module.
+
+    :param str pipeline_filepath: path to file to regard as Python module
+    :param str subtype_name: name of the target class (which must derive from
+        the base Sample class in order for it to be used), optional; if
+        unspecified, if the module defines a single subtype, then that will
+        be used; otherwise, the base Sample type will be used.
+    :return type: the imported class, defaulting to base Sample in case of
+        failure with the import or other logic
+    """
+    base_type = Sample
+
+    _, ext = os.path.splitext(pipeline_filepath)
+    if ext != ".py":
+        return base_type
+
+    try:
+        _LOGGER.debug("Attempting to import module defined by {}".
+                      format(pipeline_filepath))
+
+        # TODO: consider more fine-grained control here. What if verbose
+        # TODO: logging is only to file, not to stdout/err?
+
+        # Redirect standard streams during the import to prevent noisy
+        # error messaging in the shell that may distract or confuse a user.
+        if _LOGGER.getEffectiveLevel() > logging.DEBUG:
+            with open(os.devnull, 'w') as temp_standard_streams:
+                with utils.standard_stream_redirector(temp_standard_streams):
+                    pipeline_module = utils.import_from_source(pipeline_filepath)
+        else:
+            pipeline_module = utils.import_from_source(pipeline_filepath)
+
+    except SystemExit:
+        # SystemExit would be caught as BaseException, but SystemExit is
+        # particularly suggestive of an a script without a conditional
+        # check on __main__, and as such warrant a tailored message.
+        _LOGGER.warn("'%s' appears to attempt to run on import; "
+                     "does it lack a conditional on '__main__'? "
+                     "Using base type: %s",
+                     pipeline_filepath, base_type.__name__)
+        return base_type
+
+    except (BaseException, Exception) as e:
+        _LOGGER.warn("Using base %s because of failure in attempt to "
+                     "import pipeline module '%s': %r",
+                     base_type.__name__, pipeline_filepath, e)
+        return base_type
+
+    else:
+        _LOGGER.debug("Successfully imported pipeline module '%s', "
+                      "naming it '%s'", pipeline_filepath,
+                      pipeline_module.__name__)
+
+    def class_names(cs):
+        return ", ".join([c.__name__ for c in cs])
+
+    # Find classes from pipeline module and determine which derive from Sample.
+    classes = _fetch_classes(pipeline_module)
+    _LOGGER.debug("Found %d classes: %s", len(classes), class_names(classes))
+
+    # Base Sample could be imported; we want the true subtypes.
+    proper_subtypes = _proper_subtypes(classes, base_type)
+    _LOGGER.debug("%d proper %s subtype(s): %s", len(proper_subtypes),
+                  base_type.__name__, class_names(proper_subtypes))
+
+    # Determine course of action based on subtype request and number found.
+    if not subtype_name:
+        _LOGGER.debug("No specific subtype is requested from '%s'",
+                      pipeline_filepath)
+        if len(proper_subtypes) == 1:
+            # No specific request and single subtype --> use single subtype.
+            subtype = proper_subtypes[0]
+            _LOGGER.debug("Single %s subtype found in '%s': '%s'",
+                          base_type.__name__, pipeline_filepath,
+                          subtype.__name__)
+            return subtype
+        else:
+            # We can't arbitrarily select from among 0 or multiple subtypes.
+            # Note that this text is used in the tests, as validation of which
+            # branch of the code in this function is being hit in order to
+            # return the base Sample type. If it changes, the corresponding
+            # tests will also need to change.
+            _LOGGER.debug("%s subtype cannot be selected from %d found in "
+                          "'%s'; using base type", base_type.__name__,
+                          len(proper_subtypes), pipeline_filepath)
+            return base_type
+    else:
+        # Specific subtype request --> look for match.
+        for st in proper_subtypes:
+            if st.__name__ == subtype_name:
+                _LOGGER.debug("Successfully imported %s from '%s'",
+                              subtype_name, pipeline_filepath)
+                return st
+        raise ValueError(
+                "'{}' matches none of the {} {} subtype(s) defined "
+                "in '{}': {}".format(subtype_name, len(proper_subtypes),
+                                     base_type.__name__, pipeline_filepath,
+                                     class_names(proper_subtypes)))
+
+
+
+def _fetch_classes(mod):
+    """ Return the classes defined in a module. """
+    try:
+        _, classes = zip(*inspect.getmembers(
+                mod, lambda o: inspect.isclass(o)))
+    except ValueError:
+        return []
+    return list(classes)
+
+
+
+def _proper_subtypes(types, supertype):
+    """ Determine the proper subtypes of a supertype. """
+    return list(filter(
+            lambda t: issubclass(t, supertype) and t != supertype, types))
